django_ss/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-crispy-forms==1.14.0
Pillow==10.1.0
sqlparse==0.4.4
